Move each project deployer into a sibling folder for consistent naming
