+++ deploy/partybid/abis/PartyBid.json
