+++ deploy/partybid/configs/mainnet.json
