+++ deploy/partybuy/verify.js
@@ -1,7 +1,7 @@
 const fs = require("fs");
 const dotenv = require('dotenv');
 dotenv.config();
-const {getDeployedAddresses} = require("./helpers");
+const {getDeployedAddresses} = require("../helpers");
 
 async function verify() {
     // load .env
@@ -24,7 +24,7 @@ async function verify() {
     }
 
     // load deployed contracts
-    const {contractAddresses} = getDeployedAddresses(CHAIN_NAME);
+    const {contractAddresses} = getDeployedAddresses('partybuy', CHAIN_NAME);
     const {partyBuyFactory, partyBuyLogic, allowList} = contractAddresses;
 
     console.log(`Verifying ${CHAIN_NAME}`);
